refactor elytra into being just a process
